работа в каналах, возможность публиковать с бота в канал

+++ middlewares/antispam.py
@@ -6,9 +6,9 @@ class CheckTypeMessage(BaseMiddleware):
 
     async def on_pre_process_update(self, update: types.Update, data: dict):
         '''
-        Ограничение для юзеров, чтобы не слали в группу картинки/аудио/стикеры
+        Ограничение для юзеров, чтобы не слали в группу аудио/стикеры
         '''
         if update.message:
-            if not update.message.text:
+            if not update.message.text and not update.message.photo:
                 await update.message.delete()
                 return
